requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
curlify==2.2.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
pydantic==1.10.9
pytest==7.4.2
python-dotenv==1.0.0
requests==2.32.0
selene==2.0.0rc4
selenium==4.14.0
urllib3==2.2.2
webdriver-manager==3.8.6